Add a timeout for scan-files (#107)
This PR adds a timeout for the virus scan. This will protect Notify so it can continue to send in the event that scan-files goes down or is taking a really long time. If the virus scan has not completed by SCAN_TIMEOUT_SECONDS (currently set to 10 minutes) after the file object was created in S3, then return a 200 code along with the status scan_timed_out so the email attachment can be sent anyway.